refactor(dashboard): extract SupervisorDiagnosticsPanel to separate file
Move SupervisorDiagnosticsPanel and its helpers (registryStateBadge,
CrashCohortBar, SpEventsPanel) from keeper-detail.ts to
keeper-supervisor-diagnostics.ts.

Also remove duplicate created_at display (now in header, was also
in profile timestamps section).

keeper-detail.ts: 656 → 515 lines (-141).
